+++ src/components/ComponentWrapper/ComponentWrapper.tsx
@@ -46,12 +46,20 @@ export const ComponentWrapper: React.FC<React.PropsWithChildren<any>> = chakra(
       const hasChildren = !!children;
       const styles = useMultiStyleConfig("ComponentWrapper", { hasChildren });
       const footnote = isInvalid ? invalidText : helperText;
-      const finalHeadingText = useDSHeading({ id, title: headingText });
+      const finalHeadingText = useDSHeading({
+        id,
+        title: headingText,
+        additionalStyles: {
+          mb: "heading.default",
+        },
+      });
 
       return (
         <Box id={`${id}-wrapper`} ref={ref} __css={styles} {...rest}>
           {finalHeadingText}
-          {descriptionText && <Text>{descriptionText}</Text>}
+          {descriptionText && (
+            <Text mb="paragraph.default">{descriptionText}</Text>
+          )}
           {children}
           {footnote && (
             <HelperErrorText