+++ src/components/SocialMediaLinks/SocialMediaLinks.stories.tsx
@@ -231,6 +231,7 @@ export const LayoutVariations: Story = {
         <Heading
           id="heading-row-layout"
           level="h4"
+          mb="s"
           size="heading6"
           text="Row (Default)"
         />
@@ -240,6 +241,7 @@ export const LayoutVariations: Story = {
         <Heading
           id="heading-column-layout"
           level="h4"
+          mb="s"
           size="heading6"
           text="Column"
         />
@@ -256,6 +258,7 @@ export const SizeVariations: Story = {
         <Heading
           id="heading-small-size"
           level="h4"
+          mb="s"
           size="heading6"
           text="Small (Default)"
         />
@@ -265,6 +268,7 @@ export const SizeVariations: Story = {
         <Heading
           id="heading-medium-size"
           level="h4"
+          mb="s"
           size="heading6"
           text="Medium"
         />
@@ -274,6 +278,7 @@ export const SizeVariations: Story = {
         <Heading
           id="heading-large-size"
           level="h4"
+          mb="s"
           size="heading6"
           text="Large"
         />
@@ -290,6 +295,7 @@ export const ColorVariations: Story = {
         <Heading
           id="heading-textdefault-color"
           level="h4"
+          mb="s"
           size="heading6"
           text="Site Default"
         />
@@ -299,6 +305,7 @@ export const ColorVariations: Story = {
         <Heading
           id="heading-links-color"
           level="h4"
+          mb="s"
           size="heading6"
           text="Link"
         />
@@ -308,6 +315,7 @@ export const ColorVariations: Story = {
         <Heading
           id="heading-textinverse-color"
           level="h4"
+          mb="s"
           size="heading6"
           text="textInverse"
         />
@@ -333,6 +341,7 @@ export const BorderVariations: Story = {
         <Heading
           id="heading-border-none"
           level="h4"
+          mb="s"
           size="heading6"
           text="None (Default)"
         />
@@ -342,6 +351,7 @@ export const BorderVariations: Story = {
         <Heading
           id="heading-border-straight"
           level="h4"
+          mb="s"
           size="heading6"
           text="Straight"
         />
@@ -351,6 +361,7 @@ export const BorderVariations: Story = {
         <Heading
           id="heading-border-circular"
           level="h4"
+          mb="s"
           size="heading6"
           text="Circular"
         />
@@ -370,13 +381,20 @@ export const LabelVariations: Story = {
         <Heading
           id="heading-nolabels"
           level="h4"
+          mb="s"
           size="heading6"
           text="False (Default)"
         />
         <SocialMediaLinks showLabels={false} linksData={exampleTwoLinkData} />
       </Box>
       <Box>
-        <Heading id="heading-labels" level="h4" size="heading6" text="True" />
+        <Heading
+          id="heading-labels"
+          level="h4"
+          mb="s"
+          size="heading6"
+          text="True"
+        />
         <SocialMediaLinks showLabels={true} linksData={exampleTwoLinkData} />
       </Box>
     </VStack>