+++ src/components/Table/Table.stories.tsx
@@ -207,7 +207,7 @@ export const FontSize: Story = {
     <>
       <Stack direction="column" gap="l">
         <div>
-          <Heading level="h4" size="heading6">
+          <Heading level="h4" mb="s" size="heading6">
             desktop.body.body1
           </Heading>
           <Table
@@ -219,7 +219,7 @@ export const FontSize: Story = {
           />
         </div>
         <div>
-          <Heading level="h4" size="heading6">
+          <Heading level="h4" mb="s" size="heading6">
             desktop.body.body2
           </Heading>
           <Table
@@ -231,7 +231,7 @@ export const FontSize: Story = {
           />
         </div>
         <div>
-          <Heading level="h4" size="heading6">
+          <Heading level="h4" mb="s" size="heading6">
             desktop.caption
           </Heading>
           <Table
@@ -304,7 +304,7 @@ export const CustomColumnStyles: Story = {
   render: () => (
     <Stack spacing="xl">
       <div>
-        <Heading level="h4" size="heading6">
+        <Heading level="h4" mb="s" size="heading6">
           Without Row Headers
         </Heading>
         <Table
@@ -317,7 +317,7 @@ export const CustomColumnStyles: Story = {
         />
       </div>
       <div>
-        <Heading level="h4" size="heading6">
+        <Heading level="h4" mb="s" size="heading6">
           With Row Headers
         </Heading>
         <Table
@@ -331,7 +331,7 @@ export const CustomColumnStyles: Story = {
         />
       </div>
       <div>
-        <Heading level="h4" size="heading6">
+        <Heading level="h4" mb="s" size="heading6">
           With Empty Style Object
         </Heading>
         <Table
@@ -351,7 +351,7 @@ export const HorizontalScrolling: Story = {
   render: () => (
     <Stack spacing="xl">
       <div>
-        <Heading level="h4" size="heading5">
+        <Heading level="h4" mb="s" size="heading6">
           Without Row Headers
         </Heading>
         <Table
@@ -366,7 +366,7 @@ export const HorizontalScrolling: Story = {
         />
       </div>
       <div>
-        <Heading level="h4" size="heading5">
+        <Heading level="h4" mb="s" size="heading6">
           With Row Headers
         </Heading>
         <Table